[V8] Add a "fallback" mode for named property interceptors
By default interceptors are called before the normal property
resolution on objects.  When an interceptor is installed as a
"fallback" interceptor, it is only called if the object doesn't
already have the property.

In the case of a global object having an fallback interceptor,
the interceptor is not invoked at all for var or function
declarations.

Change-Id: Iac09d4243b0407a7b3e9ba9149f8d447b7af7766
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>