[V8] Introduce a QML compilation mode
In QML mode, there is a second global object - known as the QML
global object.  During property resolution, if a property is not
present on the JS global object, it is resolved on the QML global
object.

This global object behavior is only enabled if a script is being
compiled in QML mode.  The object to use as the QML global object
is passed as a parameter to the Script::Run() method.  Any function
closures etc. created during the run will retain a reference to this
object, so different objects can be passed in different script
runs.

Change-Id: Id3753db6f5723f6d624bd497d6f606244a4ff09a
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>