[V8] Add custom object compare callback
A global custom object comparison callback can be set with:
    V8::SetUserObjectComparisonCallbackFunction()
When two JSObjects are compared (== or !=), if either one has
the MarkAsUseUserObjectComparison() bit set, the custom comparison
callback is invoked to do the actual comparison.

This is useful when you have "value" objects that you want to
compare as equal, even though they are actually different JS object
instances.

Change-Id: I5c031b4d90c8a77f4c4ffa38270976397322db76
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>