[V8] Add support for QNX platform
This submission is based on code originally contributed by Jeff Rogers <jrogers@rim.com> with permission from Research in
Motion.

We have tried to make this change as unobtrusive as possible. For example in the atomic_ops.h header we have added a
conditional include for a qnx-specific atomicops_internals_arm_qnx.h as the existing atomicops_internals_arm_gcc.h is
actually Linux-specific despite the name of the file. In the longer term it would make sense to use an asm implementation
that is only dependent upon the CPU rather than OS-specific system calls. See qtbase/src/corelib/arch/qatomic_armv*.h
for a good reference.

Here we decided not to risk breaking existing platforms for the initial introduction of this new platform.

The QNX platform integration is built on top of the posix base.

Change-Id: If0bdf3ea3a6888472e6389ab6e74c76511e94093
Reviewed-by: Sean Harmer <sean.harmer@kdab.com>
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>