Disable snapshots
Back when v8 lived in qtbase, "cross_compile" would evaluate to
true when cross-compiling. But now, inside the qtjsbackend module,
"cross_compile" seems to evaluate to false always, which means
that snapshots are erroneously turned on when building for arm
(snapshots can't be built for arm due to QTBUG-23447).

Disable snapshots until the cause of this issue has been found.

Change-Id: If0c87ec82bf41e32bf4974b10028dcb7fa42b328
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>