Don't pass CAN_USE_VFP_INSTRUCTIONS to V8.
V8 is already capable of auto-detecting VFP3 support at runtime.
The only reason to pass this flag is for the emulator when
creating snapshots.

Currently, there is no way to figure if a target is VFP3 enabled
or not. There are no compiler flags to detect this automatically
either, so we have to add an option to qtbase/configure.

https://bugreports.qt-project.org/browse/QTBUG-24760 tracks
the addition of the option.

Change-Id: I3482ae483b925d5f153a1631cc04a52ecab72f4e
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>