mips: Add the qml mode to the lithium code generation as well
When adding qml mode to the MIPS backend, the MIPS JIT didn't have
support for lithium yet. On the upgrade to 3.10.1 MIPS gained this
support but the qml mode was not patched into the lithium code. Add
the qml mode to lithium based on the patch for the ARM jit.

Change-Id: Ia0aba2fd8b7533f38e864b54aa1d4df1e02f26c7
Reviewed-by: Kent Hansen <kent.hansen@nokia.com>
Reviewed-by: Girish Ramakrishnan <girish.1.ramakrishnan@nokia.com>