Always initialise symbol_id for ASCII symbols
Commit 30662b5d64ff633f57d89290da2a85507d561187 dropped the default
initialisation of the symbol_id for ASCII characters. This leads to
problems when trying to access the symbol_id when the symbol was not
instantiated via AsciiSymbolKey (e.g. via SubStringAsciiSymbolKey).

Additionally, AsciiSymbolKey does not guarantee the presence of a
symbol_id in case the SMI value space has been exhausted.

Change-Id: I778bc5cc3cee80038a4ca6606bfcf4423d569132
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>